+++ packages/codemod/src/codemods/jest-to-vitest-mock-factories.ts
@@ -1,5 +1,5 @@
 import type { Edit } from '@ast-grep/napi';
-import { parseTsx } from './_shared/index.js';
+import { parseTsx } from '../utils/parse.js';
 import type { Transform } from '../utils/types.js';
 
 /**
