+++ packages/codemod/src/utils/parse.ts
@@ -1,0 +1,3 @@
+import { parseAsync, Lang } from '@ast-grep/napi';
+
+export const parseTsx = (source: string) => parseAsync(Lang.Tsx, source);
